# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "BeachCrestApp",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Default server"
        }
    ],
    "paths": {
        "/api/place": {
            "get": {
                "tags": [
                    "Place"
                ],
                "summary": "Get list place",
                "description": "Get list place",
                "operationId": "index",
                "responses": {
                    "200": {
                        "description": "Thành công"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
        "/api/user": {
            "get": {
                "tags": [
                    "User Manager"
                ],
                "summary": "Get list user",
                "description": "Nhận thông tin của user",
                "operationId": "index",
                "responses": {
                    "200": {
                        "description": "Thành công"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
        "/api/auth/register": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Register",
                "description": "Register a new user and send notification mail",
                "operationId": "register",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Name",
                                        "type": "string"
                                    },
                                    "email": {
                                        "description": "Email",
                                        "type": "string"
                                    },
                                    "password": {
                                        "description": "Password",
                                        "type": "string",
                                        "format": "password"
                                    },
                                    "password_confirmation": {
                                        "description": "Confirm password",
                                        "type": "string",
                                        "format": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "422": {
                        "description": "Invalid input or email taken"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
        "/api/auth/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Login",
                "description": "Login an user",
                "operationId": "login",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "description": "Email",
                                        "type": "string"
                                    },
                                    "password": {
                                        "description": "Password",
                                        "type": "string",
                                        "format": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "422": {
                        "description": "Invalid input"
                    },
                    "403": {
                        "description": "Wrong combination of email and password or email not verified"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
        "/api/auth/getUser": {
            "get": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Get user",
                "description": "Retrieve information from current user",
                "operationId": "getUser",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
        "/api/auth/register/activate/{token}": {
            "get": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Activate user",
                "description": "Activate an registered user",
                "operationId": "activateUser",
                "parameters": [
                    {
                        "name": "token",
                        "in": "path",
                        "description": "User activating token (should be included in the verification mail)",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Invalid token"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        }
    }
}